feat: overwrite all Superset assets when initializing OARS
To do this, we're importing each individual asset type using the
"overwrite" flag. To do this, we need to add a metadata.yaml file to the
imported zip file which agrees with the imported asset type, so that is
being generated by the script instead of being stored in git.

Adds docs to explain how to maintain customizations to the
OARS-supported assets, how to share your custom assets with the
community, and how to contribute changes to OARS-maintained assets.