refactor: use course_key index in dashboard queries
This change brings back virtual datasets so that we can use the
course_name dashboard filter to leverage the course_key index in our
models. There doesn't appear to be a straightforward way to allow users
to pick from course names in the dashboard while using the course key
in our queries, so virtual datasets are the only way for us to leverage
the course_key index.