feat: remove default filters in tabs
Superset has a known bug, fixed in version 2.1.0, where dashboards will
fail to load if:
1. There are filters out of scope of the default view
2. Those filters require a default value

This change removes the default filters until Aspects is upgrade to use
version 2.1.0.

This change also includes the language code in the default text of a
translated asset. This would make it easier to identify the canonical
version of assets, such as the instructor dashboard, that should be
modified and exported when making changes.