feat: add top-level MV for grading events
This adds a top-level materialized view for grading events. This
includes course, subsection, and problem grades. Please note that the
problem grading events are also included in the problem_events
materialized view. However, it feels more appropriate to create a new MV
and duplicate these events than to try and fit the grading events into
the problem interaction MV or include the problem interaction models
whenever we want to do analyses on grades.