feat: add charts for course and problem grades
This adds two charts to the instructor dashboard along with a dataset
backing each chart. The course and problem grade datasets are configured
to return no results if there is no course or problem selected,
respectively. This is a slight break from the existing configuration,
where other charts will still show results but without distinguishing
between which course and/or problem the results are associated with. I
feel this new approach would lead to less confusion if adopted by the
other charts on each tab.