feat: add chart for full and partial video views
This adds a chart to show the number of full and partial views of course
videos. This builds on the watched video segments query and compares the
number of segments viewed by a learner to the expected number of
segments in the video. The number of segments is calculated by first
taking ten seconds off the video duration and then calculating the
number of segments based on that shortened duration. This is to account
for learners who, say, watch the entire video but skip the end credits.