feat: add video plays table to course dashboard
This change adds a table to the video engagement tab of the course
dashboard. This table shows the number of views each video has, along
with the videos section and subsection name. This is to give dashboard
viewers more context as to which videos are included in the accompanying
"video views per section/subsection" graph.