fix: Operator dash user counts, performance
Don't use the user_pii table for user counts, since it may not be populated. Change this chart to be a count of active enrollments, which people have said is more useful.

Also changed the time range for operator dash to 90 days to improve performance of the actors / events graphs.