Altinity / clustering fixes (#767)
These changes should fix the remaining issues with running Aspects in a ClickHouse Cluster such as in Altinity or using their k8s operator. It assumes you can run DDL on one node in a cluster, and that the `user_defined_zookeeper_path` ClickHouse setting is set so that user defined functions don't specifically need to be created `ON CLUSTER`.