fix: Set Clickhouse data dir permissions
On the Sumac sandbox we've seen errors regarding the ability of
ClickHouse to write to its bind mounted data directory. This hasn't
seemed to have impacted usability, and hasn't appeared elsewhere that we
have seen. This fix was suggested by Regis, and doesn't seem to break
anything in local testing but we have no ability to reproduce the issue
so it's speculative.